Refactor MSAA resolve with relocated SharedVaos and structured architecture
This final commit completes the MSAA resolve pipeline improvements with a
highly optimized and structured architecture, while relocating VAO-related
types to their correct home.

- Relocated `SharedVaos`, `SharedVao`, and `WeakVao` from `VBO.h` to
  `VAO.h` to improve modularity and follow existing project patterns.
- Implemented resolve-to-FBO and fullscreen-triangle-draw for MSAA,
  bypassing direct blit limitations in GL3.3, ES3.0, and WebGL2.
- Introduced structured `SharedVaos` and `SharedMeshes` containers in
  `Legacy::Functions` for efficient resource management per context.
- Refactored `FullScreenMesh` into a simplified, non-templated class that
  shares an `EmptyVao` across multiple shader passes (Fade and Blit).
- Optimized all fullscreen draws using `gl_VertexID` to generate NDC
  coordinates directly in the vertex shader, eliminating VBO traffic.
- Switched FBO resolve filter to `GL_NEAREST` as required for WebGL2.
- Removed legacy `m_staticMeshes` and `addSharedMesh` in favor of the new
  structured containers.
- Renamed all related resources and methods for improved clarity.

These changes provide a robust, performant, and maintainable foundation
for fullscreen effects and multisampled rendering across all platforms.
Two separate shared meshes are used because they bind different shaders
(Fade vs Blit), even though they share the same underlying geometry
(EmptyVao).
